+++ lib/service_layer/command.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require 'ostruct'
-
 module ServiceLayer
   # The +Command+ contains the logic related to the command pattern. This
   # pattern is one of the bases of the service layer. It provides the logic
@@ -34,7 +32,7 @@ module ServiceLayer
   #
   #       range = action.call(
   #         today: Time.now.utc,
-  #         date: context[:date]
+  #         date: @date
   #       )
   #
   #       { range: range }
@@ -46,11 +44,7 @@ module ServiceLayer
   module Command
     # @!visibility private
     def self.included(base)
-      base.class_eval do
-        extend ClassMethods
-
-        attr_reader :context
-      end
+      base.extend ClassMethods
     end
 
     # @!classmethods
@@ -59,32 +53,30 @@ module ServiceLayer
       #
       # Delegates the initialization, and performs the service.
       #
-      # @param context [Hash]
-      def perform(**context)
-        new(**context).perform
+      # @param properties [Hash]
+      def perform(**properties)
+        new(**properties).perform
       end
     end
 
     # Creates a new object using +Command+ module.
     #
-    # @param context [Hash]
-    def initialize(**context)
-      @context = OpenStruct.new(context)
-      context_reader
+    # Creates and assigns properties dynamically.
+    #
+    # @param properties [Hash]
+    def initialize(**properties)
+      singleton_class.class_exec(properties.keys) do |fields|
+        attr_accessor(*fields)
+      end
+
+      properties.each do |field, value|
+        __send__ "#{field}=", value
+      end
     end
 
     # @abstract must implement perform method when include command pattern.
     def perform
       raise NotImplementedError.new('Service must implement a perform method')
-    end
-
-    private
-
-    def context_reader
-      context.to_h.each_key do |key|
-        define_singleton_method(key) { context[key] }
-        define_singleton_method("#{key}=") { |value| context[key] = value }
-      end
     end
   end
 end