+++ spec/service_layer/command_spec.rb
@@ -6,8 +6,12 @@ RSpec.describe ServiceLayer::Command do
   describe '.new' do
     let(:command) { described_module.new(email: 'adriensldy@gmail.com') }
 
-    it 'creates accessor for each key' do
+    it 'creates properties' do
       expect(command).to respond_to(:email, :email=)
+    end
+
+    it 'assigns properties' do
+      expect(command).to have_attributes(email: 'adriensldy@gmail.com')
     end
   end